added mock_sapling_note
