+++ zingolib/src/test_framework/mocks.rs
@@ -1,4 +1,5 @@
 //! Tools to facilitate mocks for testing
+//! This file contains mock implementations of lrz types for unit testing with.
 use zcash_primitives::transaction::TxId;
 
 macro_rules! build_method {
@@ -9,6 +10,7 @@ macro_rules! build_method {
         }
     };
 }
+pub(crate) use build_method;
 
 // Transparent Note Mocker
 use crate::wallet::notes::TransparentNote;
@@ -65,18 +67,11 @@ impl Default for TransparentNoteBuilder {
 }
 
 // Sapling Note Mocker
-pub mod sapling_note {
+mod sapling_note {
 
-    use incrementalmerkletree::Position;
     use sapling_crypto::value::NoteValue;
     use sapling_crypto::PaymentAddress;
     use sapling_crypto::Rseed;
-    use zcash_primitives::memo::Memo;
-    use zcash_primitives::transaction::TxId;
-
-    use crate::wallet::notes::SaplingNote;
-    use crate::wallet::notes::ShieldedNoteInterface;
-    use crate::wallet::traits::FromBytes;
 
     pub struct LRZSaplingNoteBuilder {
         recipient: Option<PaymentAddress>,
@@ -112,69 +107,8 @@ pub mod sapling_note {
             }
         }
     }
-    /// builds a mock transparent note after all pieces are supplied
-    pub struct SaplingNoteBuilder {
-        diversifier: Option<sapling_crypto::Diversifier>,
-        note: Option<sapling_crypto::Note>,
-        witnessed_position: Option<Option<Position>>,
-        output_index: Option<Option<u32>>,
-        nullifier: Option<Option<sapling_crypto::Nullifier>>,
-        spent: Option<Option<(TxId, u32)>>,
-        unconfirmed_spent: Option<Option<(TxId, u32)>>,
-        memo: Option<Option<Memo>>,
-        is_change: Option<bool>,
-        have_spending_key: Option<bool>,
-    }
+}
 
-    #[allow(dead_code)] //TODO:  fix this gross hack that I tossed in to silence the language-analyzer false positive
-    impl SaplingNoteBuilder {
-        pub fn new() -> Self {
-            Self::default()
-        }
-
-        // Methods to set each field
-        build_method!(diversifier, sapling_crypto::Diversifier);
-        build_method!(note, sapling_crypto::Note);
-        build_method!(witnessed_position, Option<Position>);
-        build_method!(output_index, Option<u32>);
-        build_method!(nullifier, Option<sapling_crypto::Nullifier>);
-        build_method!(spent, Option<(TxId, u32)>);
-        build_method!(unconfirmed_spent, Option<(TxId, u32)>);
-        build_method!(memo, Option<Memo>);
-        build_method!(is_change, bool);
-        build_method!(have_spending_key, bool);
-
-        // Build method
-        pub fn build(self) -> SaplingNote {
-            SaplingNote::from_parts(
-                self.diversifier.unwrap(),
-                self.note.unwrap(),
-                self.witnessed_position.unwrap(),
-                self.nullifier.unwrap(),
-                self.spent.unwrap(),
-                self.unconfirmed_spent.unwrap(),
-                self.memo.unwrap(),
-                self.is_change.unwrap(),
-                self.have_spending_key.unwrap(),
-                self.output_index.unwrap(),
-            )
-        }
-    }
-
-    impl Default for SaplingNoteBuilder {
-        fn default() -> Self {
-            SaplingNoteBuilder {
-                diversifier: Some(sapling_crypto::Diversifier([0; 11])),
-                note: Some(LRZSaplingNoteBuilder::default().build()),
-                witnessed_position: Some(Some(Position::from(0))),
-                output_index: Some(Some(0)),
-                nullifier: Some(Some(sapling_crypto::Nullifier::from_bytes([0; 32]))),
-                spent: Some(None),
-                unconfirmed_spent: Some(None),
-                memo: Some(None),
-                is_change: Some(false),
-                have_spending_key: Some(true),
-            }
-        }
-    }
+pub fn mock_sapling_crypto_note() -> sapling_crypto::Note {
+    sapling_note::LRZSaplingNoteBuilder::default().build()
 }
